+++ warnattachment@jdede.de/defaults/preferences/prefs.js
@@ -1,2 +1,5 @@
-pref("extensions.warnattachment.warn", "DOC,DOCX,PPT,PPTX,XLS,XLSX");
+pref("extensions.warnattachment.warn", "DOC,DOCX,PPT,PPTX,XLS,XLSX,PPS,PPSX,ZIP,RAR");
 pref("extensions.warnattachment.blocked", "ADE,ADP,BAT,CHM,CMD,COM,CPL,EXE,HTA,INS,ISP,JAR,JS,JSE,LIB,LNK,MDE,MSC,MSI,MSP,MST,NSH,PIF,SCR,SCT,SHB,SYS,VB,VBE,VBS,VXD,WSC,WSF,WSH");
+
+pref("extensions.warnattachment.user_warning_msg", "");
+pref("extensions.warnattachment.user_blocked_msg", "");